requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
authlib==1.3.1
    # via weaviate-client
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==44.0.0
    # via authlib
databases==0.9.0
    # via -r requirements.in
faiss-cpu==1.9.0.post1
    # via -r requirements.in
fastapi==0.115.6
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.68.1
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.68.1
    # via weaviate-client
grpcio-tools==1.68.1
    # via weaviate-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   weaviate-client
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
loguru==0.7.3
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   faiss-cpu
packaging==24.2
    # via
    #   faiss-cpu
    #   pytest
pluggy==1.5.0
    # via pytest
protobuf==5.29.2
    # via
    #   grpcio-health-checking
    #   grpcio-tools
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   -r requirements.in
    #   fastapi
    #   weaviate-client
pydantic-core==2.27.2
    # via pydantic
pytest==8.3.4
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via databases
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
    # via -r requirements.in
validators==0.34.0
    # via weaviate-client
weaviate-client==4.10.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
